third_party/blink/web_tests/external/wpt/web-animations/interfaces/DocumentOrShadowRoot/getAnimations-expected.txt

This is a testharness.js-based test.
PASS Document.getAnimations() returns an empty sequence for non-animated content
PASS Document.getAnimations() returns script-generated animations
PASS Document.getAnimations() returns script-generated animations in the order they were created
PASS Document.getAnimations() does not return a disconnected node
PASS Document.getAnimations() does not return an animation with a null target
FAIL Document.getAnimations() returns animations on elements inside same-origin iframes assert_equals: expected 1 but got 0
FAIL ShadowRoot.getAnimations() return all animations in the shadow tree div.shadowRoot.getAnimations is not a function
FAIL Document.getAnimations() does NOT return animations in shadow trees assert_array_equals: getAnimations() called on Document does not return animations from shadow trees lengths differ, expected 0 got 1
PASS Document.getAnimations() triggers a style change event
Harness: the test ran to completion.
